requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.1.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.2.1
Django==3.1
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-filter==2.4.0
django-rest-framework==0.1.0
django-rq==2.3.2
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.4.0
flake8==3.8.3
Flask==1.1.2
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pika==1.1.0
Pillow==7.2.0
pluggy==0.13.1
promise==2.3
py==1.9.0
pycodestyle==2.6.0
pydantic==1.6.1
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
rq==1.5.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
urllib3==1.25.10
Werkzeug==1.0.1
